requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-cpp-python==0.2.90
requests==2.31.0
numpy==1.26.0
prometheus-client==0.19.0
gunicorn==21.2.0
flask-cors
